requirements.txt
streamlit==1.40.0
azure-cognitiveservices-speech==1.38.0
sounddevice==0.4.6
scipy==1.14.1
python-dotenv==1.0.1
numpy==1.26.4
wavio